pyproject.toml
[tool.poetry]
name = "openrelik-worker-parser"
version = "2025.01.09"
description = "Description of the worker"
authors = ["Dan Aschwanden <daschwanden@google.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
celery = {extras = ["redis"], version = "^5.4.0"}
openrelik-worker-common = "^2024.11.27"
asn1crypto = { version = ">= 0.24.0" }
lief = { version = ">= 0.15.1" }
pyssdeep = { version = "1.0.0" }
py-tlsh = { version = ">= 4.7.2" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
